BODS-7615 Cast timetable id to bigint instead of int (#189)

+++ ingestion_pipelines/sirivm_otp_matching_function/sirivm_otp_matching_function/sql/update_otp_state.sql
@@ -7,7 +7,7 @@ SET otp_state = CASE
                 END,
     load_time_stamp = now()::timestamp(0)
 FROM (VALUES %s) AS t(timetable_id, time_difference, last_time_in_zone_utc, is_final_stop, journey_date, timestamp_after_estimate)
-WHERE u.timetable_id = t.timetable_id::int
+WHERE u.timetable_id = t.timetable_id::bigint
   AND date_of_journey = t.journey_date::date
   AND COALESCE (
               EXTRACT(epoch FROM (t.last_time_in_zone_utc::timestamp AT TIME ZONE 'UTC' - u.expected_departure_time)),
